Move Signals module under Supervisor namespace, out of Processes
As it's just for the supervisor.
